update code ksm/013/a

+++ nts.uk/uk.at/at.web/nts.uk.at.web/src/main/webapp/view/ksm/013/a/ksm013.a.vm.ts
@@ -123,31 +123,37 @@ module nts.uk.at.view.ksm013.a {
                 let self = this,
                     dfd = $.Deferred();
                 nts.uk.ui.block.grayout();
-                let param = {
-                    nurseClassificationCode: self.nurseClModel.nurseClassificationCode()
-                }
-                service.remove(param).done(() => {
-                    nts.uk.ui.dialog.info({ messageId: "Msg_16" });
-                    if (self.lstNurseCl().length == 1) {
-                        self.lstNurseCl([]);
-                        self.newCreate();
-                    } else {
-                        let indexSelected: number;
-                        for (let index: number = 0; index < self.lstNurseCl().length; index++) {
-                            if (self.lstNurseCl()[index].code == self.selectedCode()) {
-                                indexSelected = (index == self.lstNurseCl().length - 1) ? index - 1 : index;
-                                self.lstNurseCl().splice(index, 1);
-                                break;
-                            }
-                        }
-                        self.selectedCode(self.lstNurseCl()[indexSelected].code);
+                nts.uk.ui.dialog.confirm({ messageId: 'Msg_18' }).ifYes(function() {
+                    let param = {
+                        nurseClassificationCode: self.nurseClModel.nurseClassificationCode()
                     }
+                    service.remove(param).done(() => {
+                        nts.uk.ui.dialog.info({ messageId: "Msg_16" });
+                        if (self.lstNurseCl().length == 1) {
+                            self.lstNurseCl([]);
+                            self.newCreate();
+                        } else {
+                            let indexSelected: number;
+                            for (let index: number = 0; index < self.lstNurseCl().length; index++) {
+                                if (self.lstNurseCl()[index].code == self.selectedCode()) {
+                                    indexSelected = (index == self.lstNurseCl().length - 1) ? index - 1 : index;
+                                    self.lstNurseCl().splice(index, 1);
+                                    break;
+                                }
+                            }
+                            self.selectedCode(self.lstNurseCl()[indexSelected].code);
+                        }
+                        nts.uk.ui.block.clear();
+                        dfd.resolve();
+                    });
+                }).ifNo(function() {
                     nts.uk.ui.block.clear();
                     dfd.resolve();
+                    return;
                 });
-                return dfd.promise() 
+                return dfd.promise()
             }
-            
+
             private validateAll(): boolean {
                 $('#nurseClassificationCode').ntsEditor('validate');
                 $('#nurseClassificationName').ntsEditor('validate');
@@ -164,91 +170,91 @@ module nts.uk.at.view.ksm013.a {
                 $('#nurseClassificationName').ntsError('clear');
                 $('#license').ntsError('clear');
             }
-        
-     }
+
+        }
 
 
-export class NurseClassificationModel {
-    nurseClassificationCode: KnockoutObservable<string>;
-    nurseClassificationName: KnockoutObservable<string>;
-    license: KnockoutObservable<number>;
-    officeWorker: KnockoutObservable<boolean>;
-    constructor(nurseClassificationCode: string, nurseClassificationName: string, license: number, officeWorker: boolean) {
-        let self = this;
-        self.nurseClassificationCode = ko.observable(nurseClassificationCode);
-        self.nurseClassificationName = ko.observable(nurseClassificationName);
-        self.license = ko.observable(license);
-        self.officeWorker = ko.observable(officeWorker);
-        self.license.subscribe(function(codeChanged: any) {
-             if(codeChanged != 2) self.officeWorker(false);
-        });
+        export class NurseClassificationModel {
+            nurseClassificationCode: KnockoutObservable<string>;
+            nurseClassificationName: KnockoutObservable<string>;
+            license: KnockoutObservable<number>;
+            officeWorker: KnockoutObservable<boolean>;
+            constructor(nurseClassificationCode: string, nurseClassificationName: string, license: number, officeWorker: boolean) {
+                let self = this;
+                self.nurseClassificationCode = ko.observable(nurseClassificationCode);
+                self.nurseClassificationName = ko.observable(nurseClassificationName);
+                self.license = ko.observable(license);
+                self.officeWorker = ko.observable(officeWorker);
+                self.license.subscribe(function(codeChanged: any) {
+                    if (codeChanged != 2) self.officeWorker(false);
+                });
+            }
+
+            public createDataModel(data: NurseDetailClassification) {
+                let self = this;
+                self.nurseClassificationCode(data.code);
+                self.nurseClassificationName(data.name);
+                self.license(data.license);
+                self.officeWorker(data.officeWorker);
+            }
+
+            public resetModel() {
+                let self = this;
+                self.nurseClassificationCode("");
+                self.nurseClassificationName("");
+                self.license(null);
+                self.officeWorker(false);
+
+            }
+        }
+
+        export class NurseDetailClassification {
+
+            /**
+             * 看護区分コード
+             */
+            nurseClassificationCode: string;
+
+            /**
+             * 看護区分名称   
+             */
+            nurseClassificationName: string;
+
+            /**
+             *  免許区分    
+             */
+            license: number;
+
+            /**
+             *  事務的業務従事者か
+             */
+            officeWorker: boolean;
+
+            constructor(nurseClassificationCode: string, nurseClassificationName: string, license: number, officeWorker: boolean) {
+                this.nurseClassificationCode = nurseClassificationCode;
+                this.nurseClassificationName = nurseClassificationName;
+                this.license = license;
+                this.officeWorker = officeWorker;
+            }
+        }
+
+        export class NurseClassification {
+
+            /**
+             * 看護区分コード
+             */
+            code: string;
+
+            /**
+             * 看護区分名称   
+             */
+            name: string;
+
+            constructor(nurseClassificationCode: string, nurseClassificationName: string) {
+                this.code = nurseClassificationCode;
+                this.name = nurseClassificationName;
+
+            }
+        }
     }
-
-    public createDataModel(data: NurseDetailClassification) {
-        let self = this;
-        self.nurseClassificationCode(data.code);
-        self.nurseClassificationName(data.name);
-        self.license(data.license);
-        self.officeWorker(data.officeWorker);
-    }
-
-    public resetModel() {
-        let self = this;
-        self.nurseClassificationCode("");
-        self.nurseClassificationName("");
-        self.license(null);
-        self.officeWorker(false);
-
-    }
-}
-
- export class NurseDetailClassification {
-
-    /**
-     * 看護区分コード
-     */
-    nurseClassificationCode: string;
-
-    /**
-     * 看護区分名称   
-     */
-    nurseClassificationName: string;
-
-    /**
-     *  免許区分    
-     */
-    license: number;
-
-    /**
-     *  事務的業務従事者か
-     */
-    officeWorker: boolean;
-
-    constructor(nurseClassificationCode: string, nurseClassificationName: string, license: number, officeWorker: boolean) {
-        this.nurseClassificationCode = nurseClassificationCode;
-        this.nurseClassificationName = nurseClassificationName;
-        this.license = license;
-        this.officeWorker = officeWorker;
-    }
- }
-
-export class NurseClassification {
-
-    /**
-     * 看護区分コード
-     */
-    code: string;
-
-    /**
-     * 看護区分名称   
-     */
-    name: string;
-
-    constructor(nurseClassificationCode: string, nurseClassificationName: string) {
-        this.code = nurseClassificationCode;
-        this.name = nurseClassificationName;
-
-    }
-}
-}
 }
